Optimize problem 31
This brings the runtime down from ~90s to ~3s by using a hashable data structure,
frozenset, which can in turn be put into a set instead of a list making the
detection of duplicate counts O(log n) instead of O(n).

The collections.Counter class is conceptually the perfect data type to store
the counts, but the performance characteristics of set(frozenset(tuple)) is
way better and only slightly less ergonomic.