Optimize problem 31
This was not super slow, though it was taking around a second to run.
By not keeping track of all the partitions, and only counting them, it
gets a little more efficient, but as usual - converting the arguments
to hashable types and slapping on @lru_cache makes it blazingly fast.