Optimize problem 88
Like an idiot I didn't realize that I had actually made a pretty good
algorithm, I just didn't need to run it 12 000 times.

By slightly modifying it to keep track of its progress it will generate
the solution for each k with a single pass through the tree, bringing
the time down to under a second.